requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.1
django-crispy-forms==1.8.1
django-markdown==0.8.4
django-markdownx==3.0.1
django-taggit==1.2.0
idna==2.8
Markdown==3.1.1
Pillow==6.2.1
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.9
